+++ go/pkg/cas/upload_test.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 	"sync"
 	"testing"
-	"time"
 
 	"github.com/golang/protobuf/proto"
 	"github.com/google/go-cmp/cmp"
@@ -474,32 +473,4 @@ func putSymlink(t *testing.T, path, target string) {
 	if err := os.Symlink(target, path); err != nil {
 		t.Fatal(err)
 	}
-}
-
-func TestPerCallTimeout(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	ctx, cancel, withTimeout := withPerCallTimeout(ctx, time.Millisecond)
-	defer cancel()
-
-	t.Run("succeeded", func(t *testing.T) {
-		withTimeout(func() {})
-		if ctx.Err() != nil {
-			t.Fatalf("want nil, got %s", ctx.Err())
-		}
-	})
-
-	t.Run("canceled", func(t *testing.T) {
-		withTimeout(func() {
-			select {
-			case <-ctx.Done():
-			case <-time.After(time.Second):
-			}
-		})
-
-		if ctx.Err() != context.Canceled {
-			t.Fatalf("want %s, got %s", context.Canceled, ctx.Err())
-		}
-	})
 }
